# swagger/api-principal.yaml
openapi: 3.0.3
info:
  title: API Principal PhotoSpaces
  description: >-
    Esta API es ofrecida por los microservicios de la aplicación PhotoSpaces, se
    encarga de realizar las acciones, comprobaciones

    y llamadas a las APIs de los servidores de renderizado necesarias para
    satisfacer las peticiones de los clientes.
  version: '1.0'
paths:
...
  /requests:
    get:
      tags:
        - requests
      summary: Obtener peticiones de renderizado
      description: Obtener peticiones de renderizado registradas en el sistema
      operationId: getRequests
      parameters:
        - name: limit
          in: query
          description: Número máximo de peticiones que se devolverán por estado
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Requests'
        '500':
          description: Se produce un error en la consulta a la base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRequestsServerErrorInfo'
    post:
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: String con los parámetros de renderizado en formato JSON
                model:
                  type: string
                  format: binary
                  description: Fichero con la escena 3D
                dracoCompressionLevel:
                  type: integer
                  format: int32
                  description: >-
                    Si se solicita compresión con Draco, el nivel que se
                    aplicará
                email:
                  type: string
                  format: email
                  description: >-
                    Si se elige envío por correo, la dirección a la que se
                    enviará la imagen renderizada
                requestTag:
                  type: string
                  description: >-
                    Etiqueta para facilitar la identificación de la petición en
                    el sistema
              required:
                - data
      tags:
        - requests
      summary: Enviar petición
      description: Generar petición de renderizado en el sistema
      operationId: addRequest
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddRequestSuccess'
        '400':
          description: Algún parámetro no válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddRequestInvalidParametersError'
        '500':
          description: >-
            Error relacionado con la base de datos, con el servidor de
            renderizado, con la comunicación con este, con el manejo de
            ficheros, con la compresión con Draco...
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddRequestServerError'
  '/requests/{id}':
    get:
      tags:
        - requests
      summary: Obtener petición de renderizado por id
      description: Obtener petición de renderizado a partir de su oid en la base de datos
      operationId: getRequest
      parameters:
        - name: id
          in: path
          description: Id de la petición
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Id existe pero no se corresponde con una petición finalizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRequestInvalidIdError'
        '404':
          description: Id no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRequestNotFoundError'
        '500':
          description: Error en la consulta a la base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRequestServerError'
    delete:
      tags:
        - requests
      summary: Eliminar petición de renderizado
      description: >-
        Eliminar la petición de renderizado correspondiente al parámetro id del
        sistema, independientemente de su estado, abortando el procesamiento en
        el servidor correspondiente si esta se encontrara siendo procesada
      operationId: deleteRequest
      parameters:
        - name: id
          in: path
          description: Id de la petición
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: >-
            Id no es válido o se ha enviado la petición se ha enviado en un
            momento ilógico o conflictivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteRequestInvalidIdError'
        '404':
          description: No existe servidor con el id especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteRequestNotFoundError'
        '500':
          description: >-
            Error relacionado con la base de datos, con el servidor de
            renderizado, o con la comunicación con este
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteRequestServerError'
...
components:
  schemas:
...
    Requests:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
            example: 6442a50094edb1a465f5cb14
          assignedServer:
            type: string
            example: localhost
          clientIp:
            type: string
            example: 127.0.0.1
          status:
            type: string
            example: fulfilled
          parameters:
            type: object
            properties:
              resolution:
                type: string
                example: 1080p
              bloom:
                type: boolean
                example: false
              clip_end:
                type: number
                format: double
                example: 10000
              clip_start:
                type: number
                format: double
                example: 0.1
              gtao:
                type: boolean
                example: false
              lens:
                type: number
                format: double
                example: 13.048987490261585
              location:
                type: object
                properties:
                  x:
                    type: number
                    format: double
                    example: 0
                  'y':
                    type: number
                    format: double
                    example: 0
                  z:
                    type: number
                    format: double
                    example: 0
              engine:
                type: string
                example: CYCLES
              qua:
                type: object
                properties:
                  _x:
                    type: number
                    format: double
                    example: 0
                  _y:
                    type: number
                    format: double
                    example: 0
                  _z:
                    type: number
                    format: double
                    example: 0
                  _w:
                    type: number
                    format: double
                    example: 0
              ssr:
                type: boolean
                example: false
          queueStartTime:
            type: string
            format: date-time
            example: '2023-04-21T15:00:16.398Z'
          processingStartTime:
            type: string
            format: date-time
            example: '2023-04-21T15:00:16.398Z'
          processingEndTime:
            type: string
            format: date-time
            example: '2023-04-21T15:00:16.398Z'
          requestLabel:
            type: string
            example: request1
          totalBlenderTime:
            type: integer
            format: int64
            example: 55499
          nonDeletableFile:
            type: boolean
            example: false
          __v:
            type: integer
            format: int64
            example: 0
      example:
        - _id: 6442a50094edb1a465f5cc2
          assignedServer: localhost
          clientIp: 127.0.0.1
          status: fulfilled
          parameters:
            resolution: 1080p
            bloom: false
            clip_end: 10000
            clip_start: 0.1
            gtao: false
            lens: 13.048987490261585
            location:
              x: 0
              'y': 0
              z: 0
            engine: CYCLES
            qua:
              _x: 0
              _y: 0
              _z: 0
              _w: 0
            ssr: false
          queueStartTime: '2023-04-21T15:00:16.398Z'
          processingStartTime: '2023-04-21T15:00:16.398Z'
          processingEndTime: '2023-04-21T15:00:16.398Z'
          requestLabel: request1
          totalBlenderTime: 55499
          nonDeletableFile: false
          __v: 0
        - _id: 6442a50094edb1a465f5cc6
          assignedServer: Portátil
          clientIp: 192.168.1.33
          status: fulfilled
          parameters:
            resolution: 1440p
            bloom: true
            clip_end: 10000
            clip_start: 0.1
            gtao: false
            lens: 13.048987490261585
            location:
              x: 0
              'y': 0
              z: 0
            engine: EEVEE
            qua:
              _x: 0
              _y: 0
              _z: 0
              _w: 0
            ssr: false
          queueStartTime: '2023-04-21T15:00:16.398Z'
          processingStartTime: '2023-04-21T15:00:16.398Z'
          processingEndTime: '2023-04-21T15:00:16.398Z'
          requestLabel: request2
          totalBlenderTime: 67543
          nonDeletableFile: false
          __v: 0
    GetRequestsServerErrorInfo:
      type: object
      properties:
        error:
          type: string
          example: Error en la consulta a la base de datos
    Request:
      type: object
      properties:
        _id:
          type: string
          example: 6442a50094edb1a465f5cb14
        assignedServer:
          type: string
          example: localhost
        clientIp:
          type: string
          example: 127.0.0.1
        status:
          type: string
          example: fulfilled
        parameters:
          type: object
          properties:
            resolution:
              type: string
              example: 1080p
            bloom:
              type: boolean
              example: false
            clip_end:
              type: number
              format: double
              example: 10000
            clip_start:
              type: number
              format: double
              example: 0.1
            gtao:
              type: boolean
              example: false
            lens:
              type: number
              format: double
              example: 13.048987490261585
            location:
              type: object
              properties:
                x:
                  type: number
                  format: double
                  example: 0
                'y':
                  type: number
                  format: double
                  example: 0
                z:
                  type: number
                  format: double
                  example: 0
            engine:
              type: string
              example: CYCLES
            qua:
              type: object
              properties:
                _x:
                  type: number
                  format: double
                  example: 0
                _y:
                  type: number
                  format: double
                  example: 0
                _z:
                  type: number
                  format: double
                  example: 0
                _w:
                  type: number
                  format: double
                  example: 0
            ssr:
              type: boolean
              example: false
        queueStartTime:
          type: string
          format: date-time
          example: '2023-04-21T15:00:16.398Z'
        processingStartTime:
          type: string
          format: date-time
          example: '2023-04-21T15:00:16.398Z'
        processingEndTime:
          type: string
          format: date-time
          example: '2023-04-21T15:00:16.398Z'
        requestLabel:
          type: string
          example: request1
        totalBlenderTime:
          type: integer
          format: int64
          example: 55499
        nonDeletableFile:
          type: boolean
          example: false
        __v:
          type: integer
          format: int64
          example: 0
    GetRequestInvalidIdError:
      type: object
      properties:
        error:
          type: string
          example: El parámetro id no es válido
    GetRequestNotFoundError:
      type: object
      properties:
        error:
          type: string
          example: >-
            El parámetro id no se corresponde con ninguna petición de
            renderizado almacenada en el sistema
    GetRequestServerError:
      type: object
      properties:
        error:
          type: string
          example: Error en en la consulta de la petición en la base de datos
    DeleteRequestInvalidIdError:
      type: object
      properties:
        error:
          type: string
          example: El parámetro id no es válido
    DeleteRequestNotFoundError:
      type: object
      properties:
        error:
          type: string
          example: >-
            El parámetro id no se corresponde con ninguna petición de
            renderizado almacenada en el sistema
    DeleteRequestServerError:
      type: object
      properties:
        error:
          type: string
          example: >-
            Error interno en la conexión con la base de datos tras abortar el
            procesamiento en el servidor
...
    AddRequestSuccess:
      type: object
      properties:
        requestId:
          type: string
          example: 64be2e66281bc95ec82c9bbe
        requestStatus:
          type: string
          example: processing
    AddRequestInvalidParametersError:
      type: object
      properties:
        error:
          type: string
          example: No se ha recibido archivo o este no tenía un tipo MIME válido
    AddRequestServerError:
      type: object
      properties:
        error:
          type: string
          example: Error interno durante las consultas a la base de datos
...